Fix: ejercicios M6
