+++ C6 Bucles/ejercicios/22.js
@@ -1,11 +1,8 @@
-function combine(str1, str2, str3) {
-   // Esta función debe combinar de forma alternada cada caracter de cada string.
-   // La función recibe 3 argumentos. Solo debe contabilizar aquellos que NO esten vacíos.
-   // Los strings pueden tener cualquier tamaño.
-   // EJEMPLOS
-   // combine("abc", "", "123") == "a1b2c3"
-   // combine("abc", "12345", "") == "a1b2c345"
-   // combine("abc", "12345", "67") == "a16b27c345"
+function doWhile(num) {
+  // Aumenta el valor de "num" recibido en 5 hasta un límite de 8 veces.
+  // Retorna el valor final.
+  // PISTA: Utiliza el bucle do-while.
+  // Tu código:
 }
 
-module.exports = combine;
+module.exports = doWhile;
