+++ C6 Bucles/ejercicios/23.js
@@ -1,8 +1,11 @@
-function doWhile(num) {
-  // Aumenta el valor de "num" recibido en 5 hasta un límite de 8 veces.
-  // Retorna el valor final.
-  // PISTA: Utiliza el bucle do-while.
+function esNumeroPrimo(numero) {
+  // La función recibe un argumento "numero".
+  // Determina si este corresponde a un número primo.
+  // De ser así, retorna true.
+  // De lo contrario, retorna false.
+  // IMPORTANTE: Recuerda que un número primo es aquel que
+  // solo es divisible por sí mismo y por 1.
   // Tu código:
 }
 
-module.exports = doWhile;
+module.exports = esNumeroPrimo;
