+++ C6 Bucles/tests/M6T23.test.js
@@ -1,7 +1,41 @@
-const doWhile = require('../ejercicios/23');
+const esNumeroPrimo = require("../ejercicios/23");
 
-test('Debe agregar 5 al número recibido un total de 8 veces y retornar el resultado', function () {
-  expect(doWhile(5)).toBe(45);
-  expect(doWhile(100)).toBe(140);
-  expect(doWhile(-1)).toBe(39);
+test("2 es un número primo", () => {
+  expect(esNumeroPrimo(2)).toBe(true);
+});
+
+test("3 es un número primo", () => {
+  expect(esNumeroPrimo(3)).toBe(true);
+});
+
+test("4 no es un número primo", () => {
+  expect(esNumeroPrimo(4)).toBe(false);
+});
+
+test("5 es un número primo", () => {
+  expect(esNumeroPrimo(5)).toBe(true);
+});
+
+test("6 no es un número primo", () => {
+  expect(esNumeroPrimo(6)).toBe(false);
+});
+
+test("7 es un número primo", () => {
+  expect(esNumeroPrimo(7)).toBe(true);
+});
+
+test("8 no es un número primo", () => {
+  expect(esNumeroPrimo(8)).toBe(false);
+});
+
+test("9 no es un número primo", () => {
+  expect(esNumeroPrimo(9)).toBe(false);
+});
+
+test("10 no es un número primo", () => {
+  expect(esNumeroPrimo(10)).toBe(false);
+});
+
+test("11 es un número primo", () => {
+  expect(esNumeroPrimo(11)).toBe(true);
 });
